Fix error in PlatformRunnerSettingsWindow running in 5.4
When the window is created - Unity APIs are called from the window's constructor. This is not allowed, and results in an error:

GetString is not allowed to be called from a ScriptableObject constructor (or instance field initializer), call it in OnEnable instead. Called from ScriptableObject 'PlatformRunnerSettingsWindow'.

The solution was to convert the constructor to Awake()